% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/color.R
\name{add_alpha}
\alias{add_alpha}
\title{add_alpha}
\usage{
add_alpha(colors, alpha = 1)
}
\arguments{
\item{colors}{vector of colors.}

\item{alpha}{Default is 1:No transparent.}
}
\value{
a set of transparent colours based on the alpha value specified.
}
\description{
Add transparent/alpha to a set of colours.
}
\details{
returns a set of transparent colours based on the alpha value specified.
}
\examples{
add_alpha(colors=c("black", "red", "blue"), alpha=0.5)
}

