% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utility_function.R
\name{aggregate_maxvariance}
\alias{aggregate_maxvariance}
\title{Aggregate the microarry data based on the maximum variance probe of gene probesets.}
\usage{
aggregate_maxvariance(dat, annot, symcol = "Symbol")
}
\arguments{
\item{dat}{input a expression data.frame (probes in row; samples in column).}

\item{annot}{input the annotation of the expression data (must have identical rownames).}

\item{symcol}{default is set to "Symbol" which defines the column with multiple probes
 information from the annotation.}
}
\value{
a list contains the list of "data" and "annot" with maximum variance probesets.
}
\description{
In microarray analysis, some genes are associated with multiple probes. In practice,
 sometime we need to summarize the expression in gene level (each gene has only one
 value of a given sample). This function would aggregate the microarry data based on
 the maximum variance probe of gene probesets.
}
\examples{
# Simulate gene expression data for 100 probes and 6 microarrays
# Microarray are in two groups
# First 50 probes are differentially expressed in second group
# Std deviations vary between genes with prior df=4
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*6,sd=sd),100,6))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
y[sample(seq(1,100), 10), 1] <- NA
y[sample(seq(1,100), 10), 3] <- NA
y[sample(seq(1,100), 10), 5] <- NA
set.seed(123)
annot.y <- data.frame(ID=rownames(y),
 Symbol=paste("Gene", sort(sample(seq(1, 75), 100, replace=TRUE)), sep="_"))
rownames(annot.y) <- rownames(y)
mvDat <- aggregate_maxvariance(y, annot.y, symcol="Symbol")
names(mvDat)
}
\seealso{
\code{\link{var}} which this function wraps.
}

