% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.survfit}
\alias{autoplot.survfit}
\title{Generate a ggplot2::ggplot for \code{survfit} object}
\usage{
\method{autoplot}{survfit}(object, ..., timeby = 500, xlab = "Time",
  ylab = "Survival Probability", title = "Kaplan-Meier Plot",
  titTextSize = 15, axisTitSize = 12, axisLabSize = 10, survSize = 1,
  assignCol = NULL, type = c("single", "CI", "fill"), palette = c("Dark2",
  "Set2", "Accent", "Paired", "Pastel1", "Set1", "Set3"), jitter = c("none",
  "noEvents", "all"), censShape = 3, censSize = 3, leg.x = 0.8,
  leg.y = 0.85, legend = FALSE, legLabs = NULL, legTitle = "Strata",
  legTextSize = 10, legSize = 1, alpha = 0.05, CIline = 10,
  fillLineSize = 0.05, pval = TRUE, sigP = 3, pX = 0.1, pY = 0.1,
  tabTime = c("major", "minor"), tabTitle = "Number at risk by time",
  tabTitTextSize = 15, tabLegTextSize = 5, nRiskSize = 3.5)
}
\arguments{
\item{object}{An object of class \code{survfit}}

\item{...}{Additional arguments (not implemented)}

\item{timeby}{the survival time interval}

\item{xlab}{Label for x axis on survival plot}

\item{ylab}{Label for y axis on survival plot}

\item{title}{Title for survival plot}

\item{titTextSize}{Title size for survival plot}

\item{axisTitSize}{Title size for axes}

\item{axisLabSize}{Title size for label axes}

\item{survSize}{Survival line size}

\item{assignCol}{default is NULL}

\item{type}{If \code{type="single"} (the default), plots single lines.
\cr \cr
If \code{type="CI"} will add lines indicating confidence intervals
(taken from \code{upper} and \code{lower} values of \code{survfit} object.
Higher values of \code{alpha} (tansparency) are recommended
for this, e.g. \code{alpha=0.8}.
\cr \cr
If \code{type="fill"} will add filled rectangles from the survival lines to
the confidence intervals above.}

\item{palette}{Options are taken from
\href{http://colorbrewer2.org/}{color_brewer}.
\cr \cr
\code{palette="Dark2"} (the default) is recommended for
##' \code{single} or \code{CI} plots.
\cr \cr
\code{palette="Set2"} is recommended for \code{fill} plots.}

\item{jitter}{If \code{jitter="noEvents"}, adds some random, positive noise
to survival lines with no events (i.e. all observations censored).
This will bring them just above 1 on the y-axis, making them easier
to see separately.
\cr \cr
If \code{jitter="all"} add some vertical noise to all survival lines}

\item{censShape}{Shape of marks to indicate censored onservations.
\cr Default is 3 which gives vertical ticks
\cr Use 10 for circular marks}

\item{censSize}{Size of marks to indicate censored onservations}

\item{leg.x}{input the x.axis position of the legend}

\item{leg.y}{input the y.axis position of the legend}

\item{legend}{If \code{legend=FALSE}, no legends will be produced
for the plot or table}

\item{legLabs}{These can be used to replace the names
of the strata from the fit. Should be given in the same
order as those strata}

\item{legTitle}{Title for legend}

\item{legTextSize}{Title size for legend}

\item{legSize}{Legend (key) width and height}

\item{alpha}{Alpha, transparency of lines indicating confidence intervals
or filled rectangles. Should be in range \eqn{0-1}.
Larger values e.g. \code{alpha=0.7} are recommended for confidence
intervals}

\item{CIline}{Confidence interval line type}

\item{fillLineSize}{Line size surrouding filled boxes}

\item{pval}{If \code{pval=TRUE}, adds \eqn{p} value from
log-rank test to plot}

\item{sigP}{No. of significant digits to display in \eqn{p} value.
Typically \eqn{1-3}}

\item{pX}{Location of \eqn{p} value on x axis. Should be in range of
\eqn{0 - 1}, where value is to be placed relative to the maximum observed
time. E.g. \code{pX = 0.5} will place it half-way along x-axis}

\item{pY}{Location of \eqn{p} value on y axis. Should be in range of
\eqn{0 - 1}, as above}

\item{tabTime}{If \code{tabTime="major"} this will use the major
x-axis (time) marks from the survival plot.
\cr
If \code{tabTime="minor"}, minor axis marks are used instead}

\item{tabTitle}{Table title}

\item{tabTitTextSize}{Table title text size}

\item{tabLegTextSize}{Table legend text size}

\item{nRiskSize}{No. at risk - text size}
}
\value{
A \code{list} of \code{ggplot2::ggplot} objects, with elments \code{plot},
the survial plot and \code{table} the table of events per time.
This \code{list} has the additional \code{class} of
\code{tableAndPlot}, allowing methods from \code{autoplot.tableAndPlot}.
Further modifications may be made to the objects in the list if desired.
}
\description{
Uses \code{ggplot2::ggplot2} to generate survival curves
 (Kaplan-Meier plot) and a table showing no. of events per time period
}
\details{
Generate Kaplan-Meier Plot and table using `survMisc` package in R

Generate Kaplan-Meier Plot and table using `survMisc` package in R.
}
\examples{
data(colon)
fit <- survfit(Surv(time,status)~rx, data=colon)
autoplot(fit)
autoplot(fit, type="CI", pval=TRUE, pX=0.3,
 title="Time to infection following catheter placement \\n
 by type of catheter, for dialysis patients")$plot
fit <- survfit(Surv(time,status)~1, data=colon)
autoplot(fit, legLabs="")$plot
autoplot(fit, legend=FALSE, type="fill")$plot
}
\author{
Chris Dardis. Based on existing work by
R. Saccilotto, Abhijit Dasgupta, Gil Tomas and Mark Cowley.
}

