% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utility_function.R
\name{avail_SIGinANN}
\alias{avail_SIGinANN}
\title{Check availability of a gene list in the annotation of the microarry data.}
\usage{
avail_SIGinANN(Annot, GeneList)
}
\arguments{
\item{Annot}{input the vector of characters in annotation used to match the gene list.}

\item{GeneList}{input a gene list.}
}
\value{
a summary of probes of the gene list in annotation.
}
\description{
In microarray data analysis, sometimes we only use a set of genes that PI selected for
 further analysis. We would like to first see the availability of the gene list in
 the annotation of the microarray data.
}
\examples{
x <- c(sort(sample(3:23, 100, replace=TRUE)), NA)
y <- c(sort(sample(1:20, 9)), NA)
avail_SIGinANN(x, y)
}

