% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/BI.R
\name{BI_densityplot}
\alias{BI_densityplot}
\title{Generate density-plot with Bimodality Index (BI) for microarray type of data.}
\usage{
BI_densityplot(x, G = 2, annCol = c(`1` = "Low", `2` = "High"),
  annColors = NULL, xlab = "Expression Value (log2)",
  main = "BI-Density_Plot")
}
\arguments{
\item{x}{input a numeric vector that used to calculate bimodality index (BI).}

\item{G}{input how many groups users want to separate based on mclust.}

\item{annCol}{input an information of samples that is used to combine the sample.}

\item{annColors}{input the specified colors of the sample information.}

\item{xlab}{character string specifying the label of X axis.}

\item{main}{character string specifying the graph title.}
}
\value{
a density-plot of BI value in the title of the legend with the grouping info.
}
\description{
In processing microarry data, sometime we want to find some genes with meaningful
 expression values that can separate into at least two groups. Here we wrap the
 function 'bimodalIndex' in ClassDiscovery and function 'Mclust' in mclust and plot the
 density with the grouping info.
}
\examples{
data(lungData)
BI_densityplot(x=lung.dataset["AFFX-r2-Ec-bioD-5_at",])
BI_densityplot(x=lung.dataset["AFFX-r2-Ec-bioD-5_at",],
               annCol=c("1"="AFFX-Low", "2"="AFFX-High"),
               xlab="AFFX-r2-Ec-bioD-5_at")
BI_densityplot(x=lung.dataset["AFFX-r2-Ec-bioD-5_at",],
               annCol=c("1"="AFFX-Low", "2"="AFFX-High"),
               annColors=c("AFFX-Low"="blue", "AFFX-High"="red"))
BI_densityplot(x=lung.dataset[2,], G=3, annCol=c("1"="Low", "2"="Mid", "3"="High"))
}
\seealso{
\code{\link{bimodalIndex}} which this function wraps.
}

