% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/BI.R
\name{BI_groups}
\alias{BI_groups}
\title{Separate the groups based on Bimodality Index (BI) for microarray type of data.}
\usage{
BI_groups(x, G = 2, annCol = c(`1` = "Low", `2` = "High"))
}
\arguments{
\item{x}{input a numeric vector that used to calculate bimodality index (BI).}

\item{G}{input how many groups users want to separate based on mclust.}

\item{annCol}{input an information of samples that is used to combine the sample.}
}
\value{
a character vector for the groups from 'mclust'.
}
\description{
In processing microarry data, sometime we want to find some genes with meaningful
 expression values that can separate into at least two groups. Here we wrap the
 function 'bimodalIndex' in ClassDiscovery and function 'Mclust' in mclust and return
 the the groups.
}
\examples{
data(lung.dataset)
BI_groups(x=lung.dataset["AFFX-r2-Ec-bioD-5_at",])
BI_groups(x=lung.dataset["AFFX-r2-Ec-bioD-5_at",],
           G=2, annCol=c("1"="AFFX-Low", "2"="AFFX-High"))
BI_groups(x=lung.dataset[2,], G=3, annCol=c("1"="Low", "2"="Mid", "3"="High"))
}
\seealso{
\code{\link{Mclust}} which this function wraps.
}

