% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/BI.R
\name{bimodalIndex}
\alias{bimodalIndex}
\title{Generate Bimodality Index (BI) for microarray type of data.}
\usage{
bimodalIndex(dataset, verbose = TRUE)
}
\arguments{
\item{dataset}{A matrix or data.frame, usually with columns representing samples
and rows representing genes or proteins.}

\item{verbose}{A logical value; should the function output a stream of information
 while it is working?}
}
\value{
a data frame containing six columns, with the rows corresponding to the
 rows of the original data set.  The columns contain the four parameters from the
 normal mixture model (mu1, mu2, sigma, and pi) along with the standardized
 distance 'delta' and the bimodal index (BI).
}
\description{
The bimodality index is a continuous measure of the extent to
 which a set of (univariate) data fits a two-component mixture model.
 The score is larger if the two components are balanced in size or if
 the separation between the two modes is larger.
}
\examples{
data(lung.dataset)
bi <- bimodalIndex(lung.dataset, verbose=FALSE)
summary(bi)
}
\seealso{
\code{\link{bimodalIndex}} which this function wraps.
}

