% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.PC1.scores}
\alias{calc.PC1.scores}
\title{PC1 location of samples}
\usage{
calc.PC1.scores(
  mtx,
  prb.set.df,
  prunedprobes = NULL,
  min.sd = 1,
  mingenes.pc1 = 5,
  adjust.for = NULL
)
}
\arguments{
\item{mtx}{samples by genes matrix.}

\item{prb.set.df}{a data.frame with predefined gene sets, columns are gene sets.}

\item{prunedprobes}{probes in mtx that are not used.}

\item{min.sd}{used for scaling.}

\item{mingenes.pc1}{min number of genes in a gene set for PCA.}

\item{adjust.for}{a covariates to be adjusted. Not used by me.}
}
\value{
A list of signs (location of sampmles in PCA) and signaturematrix (loadings of signature genes).
}
\description{
Mainly used for Nanostring data. Get the location of samples on PC1 calculated with signature genes in predefined gene sets.
}
\details{
Extracted and modified from Samira's codes.
}
