% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cn}
\alias{cn}
\title{Display the corner of data}
\usage{
cn(x, r = 5L, c = 5L)
}
\arguments{
\item{x}{data.}

\item{r}{number of rows to show.}

\item{c}{number of columns to show.}
}
\description{
Display the corner of data
}
