% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/extrdata.R
\docType{data}
\name{colon}
\alias{colon}
\title{a clinical information of a colon cancer dataset.}
\format{A data frame with 1858 rows and 16 variables:
\describe{
  \item{id}{Patient ID}
  \item{study}{Study}
  \item{rx}{RX}
  \item{sex}{Patient gender}
  \item{age}{Patient age}
  \item{obstruct}{obstruct}
  \item{perfor}{perfor}
  \item{adhere}{adhere}
  \item{nodes}{nodes}
  \item{status}{Vital status}
  \item{differ}{differ}
  \item{extent}{extent}
  \item{surg}{surg}
  \item{node4}{node4}
  \item{time}{time}
  \item{etype}{etype}
}}
\source{
The data is obtained from 'survival' package:
 \url{https://cran.r-project.org/web/packages/survival/index.html}
}
\usage{
colon
}
\description{
A dataset containing a data frame of 1858 samples with 16 clinical variables.
}
\keyword{datasets}

