% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corr_pcvar}
\alias{corr_pcvar}
\title{Stack the results from FactoMineR::dimdesc}
\usage{
corr_pcvar(dimdesc_res)
}
\arguments{
\item{dimdesc_res}{results from FactoMineR::dimdesc.}
}
\value{
A data.table.
}
\description{
Combined the dimdesc results as a single data.table.
}
\examples{
\dontrun{
#- Input data are samples by features.
res_pca     <- FactoMineR::PCA(decathlon[, 1:10], ncp = 5, graph = FALSE)
dimdesc_res <- FactoMineR::dimdesc(res_pca, axes = 1:3)
}
}
