% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{coxph_VecMat}
\alias{coxph_VecMat}
\title{Fit Proportional Hazards Regression Model for an expression matrix in R}
\usage{
coxph_VecMat(mat, time, event, LRT = TRUE)
}
\arguments{
\item{mat}{input a expression data.frame (probes in row; samples in column).}

\item{time}{for right censored data, this is the follow up time. For
interval data, the first argument is the starting time for the interval.}

\item{event}{The status indicator, normally 0=alive, 1=dead. Other choices
are 'TRUE'/'FALSE' ('TRUE' = death) or 1/2 (2=death).
For interval censored data, the status indicator is 0=right
censored, 1=event at 'time', 2=left censored, 3=interval
censored.  Although unusual, the event indicator can be
omitted, in which case all subjects are assumed to have an event.}
}
\value{
the correlation coefficient and p-value.
}
\description{
Fits a Cox proportional hazards regression model.  Time dependent
variables, time dependent strata, multiple events per subject, and
other extensions are incorporated using the counting process
formulation of Andersen and Gill.
}
\examples{
mat <- matrix(rnorm(400), ncol=100)
rownames(mat) <- c("A", "B","C", "D")
hmohiv<-read.table("http://www.ats.ucla.edu/stat/r/examples/asa/hmohiv.csv", sep=",", header = TRUE)
coxph_VecMat(mat=mat, time=hmohiv$time, event=hmohiv$censor)
}
\seealso{
\code{\link{coxph, Surv}} which this function wraps.
}

