% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utility_function.R
\name{create_folders}
\alias{create_folders}
\title{Create directory/folder after checking directory/folder existence using a character string}
\usage{
create_folders(fname)
}
\arguments{
\item{fname}{input the folder user would like to create.}
}
\value{
a folder with user defined name if the folder doesn't exist.
}
\description{
This function would check the directory existence. If the folder doesn't exist,
 it creates the folder with permission mode to `777`.
}
\examples{
create_folder("./test")
list.dirs("./", full.names=FALSE, recursive = FALSE)
fName <- "./test2"
create_folders(fName)
list.dirs("./", full.names=FALSE, recursive = FALSE)
}
\seealso{
\code{\link{dir.create}} which this function wraps. Also see \code{\link{create_folder}}
 for using direct path.
}

