% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/color.R
\name{diverging_palette}
\alias{diverging_palette}
\title{diverging_palette}
\usage{
diverging_palette(d = NULL, centered = TRUE, midpoint = 0,
  colors = RColorBrewer::brewer.pal(7, "PRGn"))
}
\arguments{
\item{d}{data giving the range of the palette}

\item{centered}{logical, whether to use both sides from the midpoint symmetrically}

\item{midpoint}{numeric value corresponding to the central colour}

\item{colors}{vector of colors, length must be odd}
}
\value{
a function
}
\description{
diverging colour palette function with set midpoint
}
\details{
returns a palette function that maps values to colours, with
 a midpoint (defaulting to 0) corresponding to the central colour
}
\examples{
grid.raster(diverging_palette(1:10, T, mid=2, col=c("blue", "white", "red"))(1:10))
diverging_color(1:10, T, midpoint=2, col=c("blue", "white", "red"), breaks=100)
}

