% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utility_function.R
\name{exists_robj}
\alias{exists_robj}
\title{Is an Object Defined?}
\usage{
exists_robj(object)
}
\arguments{
\item{object}{input a variable name (doesn't need to be a character string).}
}
\value{
Logical, true if and only if an object of the correct name
 and mode is found.
}
\description{
Look for an R object of the given name.
}
\examples{
x=1:10
exists_robj(x)
exists("x")  #using exists, the object must be a character string
}
\seealso{
\code{\link{exists}} which this function wraps.
}

