% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geo}
\alias{get_geo}
\title{Download and annotate GEO microarray data}
\usage{
get_geo(x, GPL, gene_anno = NULL)
}
\arguments{
\item{x}{GEO accession}

\item{GPL}{GPL id.}

\item{gene_anno}{gene annotation with three columns: probe_id, symbol, gene_id.}
}
\value{
A list of five elements if gene_anno is provided: ex_f, matrix with filted probes (20% quantile);
  ex_anno, annotated matrix; ex_g, matrix with unique symbols;
  gset, downloaded GEO data; ex_all, unfiltered matrix.
  Or a list of three elements if gene_anno is NOT provided: ex_f, matrix with filted probes (20% quantile);
  gset, downloaded GEO data; ex_all, unfiltered matrix.
}
\description{
Download data, add gene symbols, entrez ids if applicable.
}
