% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_boxplot}
\alias{gg_boxplot}
\title{Generate box-plot for QC accessment on microarray type of data.}
\usage{
gg_boxplot(dat, annCol = NULL, annColors = NULL,
  ylab = "Expression Value (log2)", outlier = TRUE, x.angle = 90,
  x.vjust = 0.5, x.size = 8, main = "Box-Plot")
}
\arguments{
\item{dat}{input a expression data.frame (probes in row; samples in column).}

\item{annCol}{input an information of samples that is used to combine the sample.}

\item{annColors}{input the specified colors of the sample information.}

\item{ylab}{character string specifying the label of y axis.}

\item{main}{character string specifying the graph title.}
}
\value{
a box-plot of microarry type of data.
}
\description{
In processing microarry data, usually we need to see how the expression values on each
 array are distributed. We put box-plots side-by-side to see the distribution within
 the dataset.
}
\examples{
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*6,sd=sd),100,6))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
y[sample(seq(1,100), 10), 1] <- NA
y[sample(seq(1,100), 10), 3] <- NA
y[sample(seq(1,100), 10), 5] <- NA
gg_boxplot(y)
gg_boxplot(y, x.angle=45, x.size=10)
gg_boxplot(y, outlier=FALSE)
annCol=c("Control","Control","Control","Tumor","Tumor","Tumor" )
annColors <- c("Control"="black", "Tumor"="red")
gg_boxplot(y, annCol=annCol)
gg_boxplot(y, annCol=annCol, annColors=annColors)
}
\seealso{
\code{\link{geom_boxplot}} which this function wraps.
}

