% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_Bum_Tab}
\alias{gg_Bum_Tab}
\title{Fit a beta-uniform mixture model to a set of p-values.}
\usage{
gg_Bum_Tab(pvals, method = "FDR", fdrs = c(0.01, 0.05, 0.1, 0.2),
  res = 100, heights = c(0.7, 0.3), main = "Bum Plot")
}
\arguments{
\item{pvals}{numeric vector containing values between '0' and '1'.}

\item{method}{either the false discovery rate (if by = 'FDR') or the posterior
probability (if by = 'EmpiricalBayes'). Default is set to 'FDR'.}

\item{fdrs}{input the false discovery rate (FDR).}

\item{res}{positive integer scalar specifying the resolution at which to plot
the fitted distribution curve.}

\item{heights}{specify the heights of Bum plot and FDR table.}

\item{main}{character string specifying the graph title.}
}
\value{
a figure with BUM plot on the top and the FDR table on the bottom.
}
\description{
In ClassComparison package in R there is a `Bum` class which is used to fit
 a beta-uniform mixture model to a set of p-values. This function wraps several
 functions, including 'Bum', 'hist', 'cutoffSignificant', and 'selectSignificant'
 into one function and presents the results in one figure.
}
\examples{
fake.data <- c(runif(700), rbeta(300, 0.3, 1))
gg_Bum_Tab(fake.data, main="", heights=c(0.75, 0.25))
}
\seealso{
\code{\link{Bum}}, \code{\link{hist}}, \code{\link{cutoffSignificant}},
 and \code{\link{selectSignificant}} which this function wraps.
}

