% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_cef_Tab}
\alias{gg_cef_Tab}
\title{Generate a plot of table shows the features whose coefficient greater than cutoffs.}
\usage{
gg_cef_Tab(cefs, annot, cutoffs = c(0.5, 0.6, 0.7, 0.8), main = "",
  footnote = "")
}
\arguments{
\item{cefs}{numeric vector containing values between '0' and '1'.}

\item{annot}{input the name of features used in the figure.}

\item{cutoffs}{input the cutoff of the p-values.}

\item{main}{character string specifying the graph title.}
}
\value{
a figure of table shows the features whose p-values less than cutoffs.
}
\description{
We like to list the features whose coefficient greater than certain cutoffs in a table.
}
\examples{
fake.data <- c(rbeta(10, 0.8, 1))
annot.name <- paste("Probe", 1:100, sep="_")
gg_cef_Tab(fake.data, annot.name)
gg_cef_Tab(fake.data, annot.name, main="Tumor vs. Normal", footnote="Without FDR correction.")
}
\seealso{
\code{\link{Bum}}, \code{\link{hist}}, \code{\link{cutoffSignificant}},
 and \code{\link{selectSignificant}} which this function wraps.
}

