% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_corplot_XY}
\alias{gg_corplot_XY}
\title{Generate correlation-plot between X and Y.}
\usage{
gg_corplot_XY(x, y, xlab = NULL, ylab = NULL, dot.size = 2)
}
\arguments{
\item{x}{input numeric vector to specify the coordinate on X-axis.}

\item{y}{input numeric vector to specify the coordinate on Y-axis.}

\item{xlab}{character string specifying the label of X axis.}

\item{ylab}{character string specifying the label of y axis.}

\item{dot.size}{input the size of the dot. Default is 2.}
}
\value{
a correlation-plot of X and Y.
}
\description{
Generate correlation-plot between X and Y, returning results for both Pearson and
 Spearman correlation.
}
\examples{
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*6,sd=sd),100,6))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
y[sample(seq(1,100), 10), 1] <- NA
y[sample(seq(1,100), 10), 3] <- NA
y[sample(seq(1,100), 10), 5] <- NA
gg_corplot_XY(x=y[,1], y=y[,2])
gg_corplot_XY(x=y[,1], y=y[,2], xlab="Random X", ylab="Random Y")
}
\seealso{
\code{\link{rcorr_XY}} which this function wraps.
}

