% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_fisher.test_Tab}
\alias{gg_fisher.test_Tab}
\title{Generate a plot of table shows the results of Fisher's Exact Test for Count Data.}
\usage{
gg_fisher.test_Tab(x, y = NULL, main = "Fisher's Exact Test for Count Data",
  show.dimnames = FALSE, ...)
}
\arguments{
\item{x}{either a two-dimensional contingency table in matrix form, or a factor object.}

\item{y}{a factor object; ignored if 'x' is a matrix.}

\item{workspace}{an integer specifying the size of the workspace used in the network algorithm.
In units of 4 bytes.  Only used for non-simulated p-values larger than 2 by 2 tables.}

\item{hybrid}{a logical. Only used for larger than 2 by 2 tables, in which cases it indicates
whether the exact probabilities (default) or a hybrid approximation thereof should be computed.
 See 'Details'.}

\item{control}{a list with named components for low level algorithm control. At present the only
one used is '"mult"', a positive integer >= 2 with default 30 used only for larger than 2 by 2
tables. This says how many times as much space should be allocated to paths as to keys: see
file 'fexact.c' in the sources of this package.}

\item{or}{the hypothesized odds ratio.  Only used in the 2 by 2 case.}

\item{alternative}{indicates the alternative hypothesis and must be one of '"two.sided"',
'"greater"' or '"less"'.  You can specify just the initial letter.  Only used in the 2 by 2 case.}

\item{conf.int}{logical indicating if a confidence interval for the odds ratio in a 2 by 2 table
should be computed (and returned).}

\item{conf.level}{confidence level for the returned confidence interval. Only used in the 2 by 2
case and if 'conf.int = TRUE'.}

\item{simulate.p.value}{a logical indicating whether to compute p-values by Monte Carlo simulation,
in larger than 2 by 2 tables.}

\item{B:}{an integer specifying the number of replicates used in the Monte Carlo test.}
}
\value{
a figure of table shows the results of Fisher's Exact Test for Count Data.
}
\description{
Performs Fisher's exact test for testing the null of independence of rows and columns
 in a contingency table with fixed marginals.
}
\examples{
TeaTasting <- matrix(c(3, 1, 1, 3),
                    nrow = 2, dimnames = list(Guess = c("Milk", "Tea"),
                    Truth = c("Milk", "Tea")))
TeaTasting
fisher.test(TeaTasting)
gg_fisher.test_Tab(TeaTasting, show.dimnames=TRUE)
Job <- matrix(c(1,2,1,0, 3,3,6,1, 10,10,14,9, 6,7,12,11), 4, 4,
             dimnames = list(income = c("< 15k", "15-25k", "25-40k", "> 40k"),
                             satisfaction = c("VeryD", "LittleD", "ModerateS", "VeryS"))
             )
Job
fisher.test(Job)
gg_fisher.test_Tab(Job)
}
\seealso{
\code{\link{fisher.test}} which this function wraps.
}

