% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_PCA}
\alias{gg_PCA}
\title{Generate principal components analysis (PCA) Plot on microarray type of data.}
\usage{
gg_PCA(x, annCol = NULL, annColors = NULL, annS = NULL,
  annShapes = NULL, main = "1st & 2nd PC")
}
\arguments{
\item{x}{input a expression data.frame (probes in row; samples in column).}

\item{annCol}{input an information of samples that is used to distinguish the
sample in colors.}

\item{annColors}{input the specified colors of the sample information.}

\item{annS}{input an information of samples that is used to distinguish the
sample in shapes.}

\item{annShapes}{input the specified shapes of the sample information.}

\item{main}{character string specifying the graph title.}
}
\value{
a density-plot of microarry type of data.
}
\description{
In processing microarry data, usually we need to see how the expression values on each
 array are separated in space.
}
\examples{
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*6,sd=sd),100,6))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
annCol=c("Control","Control","Control","Tumor","Tumor","Tumor" )
annS=c("Batch1","Batch1","Batch2","Batch2","Batch3","Batch3" )
annColors <- c("Control"="black", "Tumor"="red")
annShapes=c("Batch1"=1,"Batch2"=4,"Batch3"=6 )
gg_PCA(y)
gg_PCA(y, annCol=annCol)
gg_PCA(y, annS=annS)
gg_PCA(y, annCol=annCol, annColors=annColors)
gg_PCA(y, annS=annS, annShapes=annShapes)
gg_PCA(y, annCol=annCol, annS=annS)
}
\seealso{
\code{\link{SamplePCA}} which this function wraps.
}

