% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_pval_Tab}
\alias{gg_pval_Tab}
\title{Generate a plot of table shows the features whose p-values less than cutoffs.}
\usage{
gg_pval_Tab(pvals, annot, cutoffs = c(0.01, 0.05, 0.1), main = "",
  footnote = "")
}
\arguments{
\item{pvals}{numeric vector containing values between '0' and '1'.}

\item{annot}{input the name of features used in the figure.}

\item{cutoffs}{input the cutoff of the p-values.}

\item{main}{character string specifying the graph title.}
}
\value{
a figure of table shows the features whose p-values less than cutoffs.
}
\description{
In ClassComparison package in R there is a `Bum` class which is used to fit
 a beta-uniform mixture model to a set of p-values. However sometimes the number
 of features are limitted, we would like to expand the feature list by ignoring
 FDR correstion.
}
\examples{
fake.data <- c(rbeta(100, 0.8, 1))
annot.name <- paste("Probe", 1:100, sep="_")
gg_pval_Tab(fake.data, annot.name)
gg_pval_Tab(fake.data, annot.name, main="Tumor vs. Normal", footnote="Without FDR correction.")
}
\seealso{
\code{\link{Bum}}, \code{\link{hist}}, \code{\link{cutoffSignificant}},
 and \code{\link{selectSignificant}} which this function wraps.
}

