% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Generate Quantile-Quantile Plots.}
\usage{
gg_qqplot(vec, dot.size = 3, xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles", main = "Normal Q-Q Plot")
}
\arguments{
\item{vec}{input numeric vector.}

\item{dot.size}{input the size of the dot. Default is 3.}

\item{xlab}{character string specifying the label of X axis.}

\item{ylab}{character string specifying the label of y axis.}

\item{main}{Character giving the main title of the diagram.}
}
\value{
a correlation-plot of X and Y.
}
\description{
Generate a normal QQ plot of the values in vec using qqnorm function and adds a line
 to a "theoretical", by default normal, quantile-quantile plot which passes through
 the probs quantiles, by default the first and third quartiles.
}
\examples{
vec <- rt(200, df = 5)
gg_qqplot(vec=vec)
gg_qqplot(vec=vec, size=2)
}
\seealso{
\code{\link{qqnorm}} and \code{\link{qqline}} which this function wraps.
}

