% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{gg_venn.diagram}
\alias{gg_venn.diagram}
\title{Make a Venn Diagram with count table that each component corresponding to a separate circle in VD.}
\usage{
gg_venn.diagram(x, countFile = "vennDiag_counts.csv", ...)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding
to a separate circle in the Venn diagram}

\item{...}{A series of graphical parameters tweaking the plot. See below for details.}

\item{filename}{Default is set to NULL which returns the grid object itself.}

\item{height}{Integer giving the height of the output figure in units.}

\item{width}{Integer giving the width of the output figure in units.}

\item{resolution}{Resolution of the final figure in DPI.}

\item{units}{Size-units to use for the final figure.}

\item{compression}{What compression algorithm should be applied to the final tiff.}

\item{na}{Missing value handling method: "none", "stop", "remove".}

\item{main}{Character giving the main title of the diagram.}

\item{sub}{Character giving the subtitle of the diagram.}

\item{main.pos}{Vector of length 2 indicating (x,y) of the main title.}

\item{main.fontface}{Character giving the fontface (font style) of the main title.}

\item{main.fontfamily}{Character giving the fontfamily (font type) of the main title.}

\item{main.col}{Character giving the colour of the main title.}

\item{main.cex}{Number giving the cex (font size) of the main title.}

\item{main.just}{Vector of length 2 indicating horizontal and vertical justification of the main title.}

\item{sub.pos}{Vector of length 2 indicating (x,y) of the subtitle.}

\item{sub.fontface}{Character giving the fontface (font style) of the subtitle.}

\item{sub.fontfamily}{Character giving the fontfamily (font type) of the subtitle.}

\item{sub.col}{Character Colour of the subtitle.}

\item{sub.cex}{Number giving the cex (font size) of the subtitle.}

\item{sub.just}{Vector of length 2 indicating horizontal and vertical justification of the subtitle.}

\item{category.names}{Allow specification of category names using plotmath syntax.}

\item{force.unique}{Logical specifying whether to use only unique elements in each item of the input
list or use all elements. Defaults to FALSE}
}
\value{
A Venn Diagram with a count table that each component corresponding to a separate circle in VD.
}
\description{
This function takes a list and creates a publication-quality Venn Diagram.
}
\examples{
mylist <- list(A = 1:150, B = 121:170)
gg_venn.diagram(x=mylist,
               col = "transparent",
               fill=rainbow(2), alpha=0.5,
               cex = 1.5, cat.fontface = 4, lty =2,
               fontfamily =3,
               margin = 0.1, main="mylist",
               countFile="vennDiag_mylist.csv")
mylist <- list(A = 1:150, B = 121:170, C = 101:200)
gg_venn.diagram(x=mylist,
               col = "transparent",
               fill=rainbow(3), alpha=0.5,
               cex = 1.5, cat.fontface = 4, lty =2,
               fontfamily =3,
               margin = 0.1, main="mylist",
               countFile="vennDiag_mylist.csv")
mylist <- list(
 I = c(1:60, 61:105, 106:140, 141:160, 166:175, 176:180, 181:205, 206:220),
 IV = c(531:605, 476:530, 336:375, 376:405, 181:205, 206:220, 166:175, 176:180),
 II = c(61:105, 106:140, 181:205, 206:220, 221:285, 286:335, 336:375, 376:405),
 III = c(406:475, 286:335, 106:140, 141:160, 166:175, 181:205, 336:375, 476:530)
)
gg_venn.diagram(x=mylist,
               col = "transparent",
               fill=rainbow(4), alpha=0.5,
               cex = 1.5, cat.fontface = 4, lty =2,
               fontfamily =3,
               margin = 0.1, main="mylist",
               countFile="vennDiag_mylist.csv")
}
\seealso{
\code{\link{venn.diagram}} which this function wraps.
}

