% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david.R
\name{GO_plot_david}
\alias{GO_plot_david}
\title{GO plots for DAVID GO annotation}
\usage{
GO_plot_david(
  in_file,
  height = NULL,
  width = 7,
  thresh_p = 0.001,
  thresh_fdr = NULL,
  lollipop = TRUE,
  cex_size = 6,
  col_palette = "#0073c2ff",
  output_name = NULL,
  save_as_pdf = FALSE,
  return_plot = FALSE
)
}
\arguments{
\item{in_file}{path to the david GO results.}

\item{height}{pass to \code{"ggsave"} height.}

\item{width}{pass to \code{"ggsave"} width.}

\item{thresh_p}{P_value threshold.}

\item{thresh_fdr}{FDR threshold (default NULL and using the P value threshold).}

\item{lollipop}{whether using lollipop plots.}

\item{cex_size}{size for the row names.}

\item{col_palette}{color to be used for barplot.}

\item{output_name}{name of the output file; by default using the base name of in_file.}

\item{save_as_pdf}{whether to save as pdf.}

\item{return_plot}{whether to reture the ggplot object.}
}
\value{
Save a png/pdf file.
}
\description{
Put the GO terms in lollipop plots or barplots.
}
