% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{GSEA_heatmap}
\alias{GSEA_heatmap}
\title{Summarize the GSEA results by heatmap}
\usage{
GSEA_heatmap(
  dta,
  cate = "h",
  NES_thresh = NULL,
  only_sig = FALSE,
  return_comb = TRUE,
  row_font_size = 5,
  col_font_size = 8,
  col_order_new = NULL,
  width = 4,
  height = 6,
  output_name = "hallmark",
  color_palette = c("navy", "white", "firebrick3"),
  ...
)
}
\arguments{
\item{dta}{the input GSEA result lists.}

\item{cate}{GSEA annotation category, like h, or kegg.}

\item{NES_thresh}{filter the non-sig results, if NULL then no filtering.}

\item{only_sig}{whether only plotting significant gene sets (FDR <= .25).}

\item{return_comb}{whether return the Combined results.}

\item{row_font_size}{fontsize for row labels.}

\item{col_font_size}{fontsize for column labels.}

\item{col_order_new}{new order of samples in matrix, only works if cluster_columns is FALSE.}

\item{width}{width of the heatmap.}

\item{height}{height of the heatmap.}

\item{output_name}{the name of output file.}

\item{color_palette}{color used in heatmap.}

\item{...}{additional parameters to Heatmap.}
}
\value{
A data.table if return_comb is TRUE.
}
\description{
Plot a heatmap based on a list with different GSEA category annotation.
}
