% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{limma_anova_contrast}
\alias{limma_anova_contrast}
\title{Perform One-way ANOVA with contrast using `limma` package in R}
\usage{
limma_anova_contrast(data, fac, contrast)
}
\arguments{
\item{data}{input a expression data.frame (probes in row; samples in column).}

\item{fac}{input the sample grouping info that you want to compare.}

\item{contrast}{inpt the contrast between any two given groups in 'fac'.}
}
\value{
a list of two tables. 'dds' contains the original data used to perform the test;
 `results` contains the testing statistics.
}
\description{
Fit linear model for each gene given a series of arrays using `limma` package in R.
}
\examples{
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*9,sd=sd),100,9))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""), paste("C", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
y[51:75,7:9] <- y[51:75,7:9] + 3
y[sample(seq(1,100), 10), 1] <- NA
y[sample(seq(1,100), 10), 3] <- NA
y[sample(seq(1,100), 10), 5] <- NA
mFAC <- c(rep("Control", 3), rep("Treat1", 3), rep("Treat2", 3))
mContr <- c("Treat1-Control", "Treat2-Control", "Treat2-Treat1")
comTest <- limma_anova_contrast(data=y, fac=mFAC, contrast=mContr)
names(comTest)
comTest[[2]][1:5,]
gg_Bum_Tab(comTest[[2]][,'p.value'], main="Controls vs. Tumors", heights=c(0.75, 0.25))
}
\seealso{
\code{\link{lmFit}} which this function wraps.
}

