% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{limma_ttest}
\alias{limma_ttest}
\title{Perform Two-Sample t-tests using `limma` package in R}
\usage{
limma_ttest(data, ctlS, trtS, ctlS.name = NULL, trtS.name = NULL)
}
\arguments{
\item{data}{input a expression data.frame (probes in row; samples in column).}

\item{ctlS}{input the sample names of the control group as the reference.}

\item{trtS}{input the sample names of the treatment group.}

\item{ctlS.name}{input the name of the control group using in the comparison.}

\item{trtS.name}{input the name of the treatment group using in the comparison.}
}
\value{
a list of two tables. 'dds' contains the original data used to perform the test;
 `results` contains the testing statistics.
}
\description{
Fit linear model for each gene given a series of arrays using `limma` package in R.
}
\examples{
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- data.frame(matrix(rnorm(100*6,sd=sd),100,6))
rownames(y) <- paste("Probe",1:100, sep="_")
colnames(y) <- c(paste("A", 1:3, sep=""), paste("B", 1:3, sep=""))
y[1:50,4:6] <- y[1:50,4:6] + 2
y[sample(seq(1,100), 10), 1] <- NA
y[sample(seq(1,100), 10), 3] <- NA
y[sample(seq(1,100), 10), 5] <- NA
Controls=c("A1","A2","A3")
Tumors=c("B1","B2","B3")
comTest <- limma_ttest(y, ctlS=Controls, trtS=Tumors)
comTest2 <- limma_ttest(y, ctlS=Controls, trtS=Tumors, ctlS.name="Normal", trtS.name="Cancer")
names(comTest)
comTest[[2]][1:5,]
comTest2[[2]][1:5,]
gg_Bum_Tab(comTest[[2]][,'p.value'], main="Controls vs. Tumors", heights=c(0.75, 0.25))
}
\seealso{
\code{\link{lmFit}} which this function wraps.
}

