% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/extrdata.R
\docType{data}
\name{lung.clinical}
\alias{lung.clinical}
\title{A subset of clinical information of a lung cancer dataset.}
\format{A data frame with 444 rows and 10 variables:
\describe{
  \item{PATIENT_ID}{Patient ID}
  \item{SITE}{Tumor site}
  \item{GENDER}{Patient gender}
  \item{AGE_AT_DIAGNOSIS}{Diagnosis age}
  \item{VITAL_STATUS}{Vital status}
  \item{MONTHS_TO_LAST_CONTACT_OR_DEATH}{Moths to last contact or death}
  \item{SMOKING_HISTORY}{Smoking history}
  \item{PATHOLOGIC_N_STAGE}{Pathologic N stage}
  \item{PATHOLOGIC_T_STAGE}{Pathologic T stage}
  \item{Histologic.grade}{Histologic grade}
}}
\source{
The data is obtained from 'ClassDiscovery' package:
 \url{http://bioinformatics.mdanderson.org/main/OOMPA:Overview}
}
\usage{
lung.clinical
}
\description{
A dataset containing a data frame of 444 samples with 10 clinical variables.
}
\keyword{datasets}

