% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/extrdata.R
\docType{data}
\name{lung.dataset}
\alias{lung.dataset}
\title{A subset of expression microarray of a lung cancer dataset.}
\format{A matrix with 150 rows and 444 samples:
\describe{
  \item{row}{Probe ID}
  \item{col}{Sample ID}
}}
\source{
The data is obtained from 'ClassDiscovery' package:
 \url{http://bioinformatics.mdanderson.org/main/OOMPA:Overview}
}
\usage{
lung.dataset
}
\description{
A dataset containing a matrix of 444 samples with 150 probes.
}
\keyword{datasets}

