% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{make_unique}
\alias{make_unique}
\title{Make a vector unique.}
\usage{
make_unique(x, sep_char = "_")
}
\arguments{
\item{x}{input vector.}

\item{sep_char}{suffix.}
}
\value{
A unique vector.
}
\description{
Add the sep delim to duplicated elements of vector based on their location in the vector, starting form 1.
}
