% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchstring}
\alias{matchstring}
\title{Pattern to grep}
\usage{
matchstring(x, exact = TRUE)
}
\arguments{
\item{x}{a vector of characters to build the string.}

\item{exact}{exact match of vectors (default TRUE).}
}
\value{
A string to be used in grep or similar regex setting.
}
\description{
Build a string from vectors to grep for exact match.
}
