% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_stallion_20}
\alias{palette_stallion_20}
\title{stallion 20 color}
\format{
A 20 color palette extracted from ArchR.
}
\usage{
palette_stallion_20
}
\description{
stallion 20 color
}
\keyword{datasets}
