% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_data.R
\docType{data}
\name{palette_zeileis_28}
\alias{palette_zeileis_28}
\title{scvelo 28 color}
\format{
A color palette extracted from scvelo.
}
\usage{
palette_zeileis_28
}
\description{
scvelo 28 color
}
\keyword{datasets}
