% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_plots.R
\name{plt_stream}
\alias{plt_stream}
\title{Makes a stream plot}
\usage{
plt_stream(
  x,
  y,
  order.method = "as.is",
  center = TRUE,
  frac.rand = 0.1,
  spar = 0.2,
  xlab = "",
  ylab = "",
  col = rainbow(ncol(y)),
  ylim = NULL,
  border = NULL,
  lwd = 1,
  ...
)
}
\arguments{
\item{x}{a vector of value to be used in x axis.}

\item{y}{a matrix of m * n, while m equals to the length of x. The columns are the value to be plot.}

\item{order.method}{a value of c('as.is', 'max', 'first').}

\item{center}{if TRUE, the stacked polygons will be centered so that the middle, i.e. baseline ('g0'), of the stream is approximately equal to zero.}

\item{frac.rand}{fraction of the overall data 'stream' range used to define the range of random wiggle (uniform distrubution) to be added to the baseline g0.}

\item{spar}{setting for smooth.spline function to make a smoothed version of baseline g0.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{col}{fill colors for polygons corresponding to y columns (will recycle).}

\item{ylim}{limts for y axis.}

\item{border}{border colors for polygons corresponding to y columns (will recycle) (see ?polygon for details).}

\item{lwd}{border line width for polygons corresponding to y columns (will recycle).}

\item{...}{other plot arguments.}
}
\value{
A plot.
}
\description{
Make a stream plot with value on a matrix.
}
\details{
In order.method, as.is, plot in order of y column; max, plot in order of when each y series reaches maximum value;
  first, plot in order of when each y series first value > 0.

  Centering is done before the addition of random wiggle to the baseline.
}
\examples{
\dontrun{
set.seed(1)
m <- 500
n <- 50
x <- seq(m)
y <- matrix(0, nrow=m, ncol=n)
colnames(y) <- seq(n)
for(i in seq(ncol(y))) {
    mu <- runif(1, min=0.25*m, max=0.75*m)
    SD <- runif(1, min=5, max=30)
    TMP <- rnorm(1000, mean=mu, sd=SD)
    HIST <- hist(TMP, breaks=c(0,x), plot=FALSE)
    fit <- smooth.spline(HIST$counts ~ HIST$mids)
    y[,i] <- fit$y
}

y <- replace(y, y<0.01, 0)

COLS <- rainbow(ncol(y))

plt.stream(x,y, axes=FALSE, xlim=c(100, 400), xaxs="i", center=TRUE,
           spar=0.2, frac.rand=0.1, col=COLS, border=1, lwd=0.1)
}
}
\references{
Stackover thread, https://stackoverflow.com/a/13087137/349054
}
