% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{rcorr_VecMat}
\alias{rcorr_VecMat}
\title{Perform Correlation tests between each row in a matrix and a vector using `Hmisc` package in R}
\usage{
rcorr_VecMat(vec, mat, method = c("pearson", "spearman"))
}
\arguments{
\item{vec}{a numeric vector which will be concatenated to the row of 'mat'.}

\item{mat}{a numeric matrix with at least 5 rows and at least 2 columns.
For print, x is an object produced by rcorr.}

\item{method}{specifies the type of correlations to compute. Spearman correlations
are the Pearson linear correlations computed on the ranks of non-missing elements,
using midranks for ties.}
}
\value{
the correlation coefficient and p-value.
}
\description{
rcorr Computes a matrix of Pearson's r or Spearman's rho rank correlation
 coefficients for all possible pairs of columns of a matrix. Missing values
 are deleted in pairs rather than deleting all rows of x having any missing
 variables. Ranks are computed using efficient algorithms, using midranks for
 ties.
}
\examples{
mat <- matrix(rnorm(200), ncol=50)
rownames(mat) <- c("A", "B","C", "D")
vec <- rnorm(50)
rcorr_VecMat(vec=vec, mat=mat, method='pearson')
rcorr_VecMat(vec=vec, mat=mat, method='spearman')
}
\seealso{
\code{\link{rcorr}} which this function wraps.
}

