% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{rcorr_XY}
\alias{rcorr_XY}
\title{Perform Correlation tests using `Hmisc` package in R}
\usage{
rcorr_XY(x, y, method = c("pearson", "spearman"), showname = NULL)
}
\arguments{
\item{x}{a numeric matrix with at least 5 rows and at least 2 columns
(if y is absent). For print, x is an object produced by rcorr.}

\item{y}{a numeric vector or matrix which will be concatenated to x. If y is
omitted for rcorr, x must be a matrix.}

\item{method}{specifies the type of correlations to compute. Spearman correlations
are the Pearson linear correlations computed on the ranks of non-missing elements,
using midranks for ties.}

\item{showname}{default is NULL. It will show the name of the method selected,
 otherwise the character would be used to assign the rowname of the result.}
}
\value{
the correlation coefficient and p-value.
}
\description{
rcorr Computes a matrix of Pearson's r or Spearman's rho rank correlation
 coefficients for all possible pairs of columns of a matrix. Missing values
 are deleted in pairs rather than deleting all rows of x having any missing
 variables. Ranks are computed using efficient algorithms, using midranks for
 ties.
}
\examples{
rcorr_XY(x=mtcars[,1], y=mtcars[,2], method='pearson')
rcorr_XY(x=mtcars[,1], y=mtcars[,2], method='spearman', showname="IC50")
}
\seealso{
\code{\link{rcorr}} which this function wraps.
}

