% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{readrsem}
\alias{readrsem}
\title{Read RSEM result files}
\usage{
readrsem(inputfiles, is_gene = TRUE)
}
\arguments{
\item{inputfiles}{path to input files.}

\item{is_gene}{whethe the input are genes (default) or transcript.}
}
\value{
A list of three type of data, each is a data.table of all samples.
}
\description{
Read the RSEM output files, either gene results or transcript results, to a
a list with count, TPM, fpkm.
}
\examples{
\dontrun{
inputfiles <- dir("/results", pattern=".*.genes.results$", full.names = TRUE, recursive = TRUE)
res <- readrsem(inputfiles)
countdata <- res$count_data
tpmdata <- res$tpm_data
fpkmdata <- res$fpkm_data
}
}
