% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{rs2}
\alias{rs2}
\title{Split a string into two part from right end}
\usage{
rs2(x, s, n)
}
\arguments{
\item{x}{a string to be splitted.}

\item{s}{the separator}

\item{n}{the right location of s.}
}
\value{
A vector with splitted parts.
}
\description{
Split a string from the righ end and return the result vector.
}
\examples{
\dontrun{
rs2("A_B_C_D_E", "_", 3)
# "A_B", "C_D_E"
}
}
