% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utility_function.R
\name{show_robj}
\alias{show_robj}
\title{Show a list of names of objects from a R file/RData.}
\usage{
show_robj(RData)
}
\arguments{
\item{RData}{input a R file/RData containing objects.}
}
\value{
a vector of character strings giving the names of
 the objects in a R file/RData.
}
\description{
This function allows you to list the names of objects from
 a R file/RData
}
\examples{
}
\seealso{
\code{\link{objects}} which this function wraps.
}

