% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{stats_chisq.test}
\alias{stats_chisq.test}
\title{Perform Pearson's Chi-squared Test between each row in a matrix and a vector using `stats` package in R}
\usage{
stats_chisq.test(data, group, rowlevel, ...)
}
\arguments{
\item{data}{a matrix with at least 2 outcomes (0/1; Mut/WT, etc).}

\item{group}{a vector of factor which will be concatenated to the row of 'data'.}

\item{rowlevel}{specifies the order of factor use in a contingency table.}

\item{...}{other parameters in chisq.test function.}
}
\value{
the contingency table and p-value.
}
\description{
'chisq.test' performs chi-squared contingency table tests and goodness-of-fit tests.
}
\examples{
m0 <- matrix(0, 5, 30)
dat <- apply(m0, c(1,2), function(x) sample(c(0,1),1))
rownames(dat) <- paste("R", 1:5, sep="")
colnames(dat) <- paste("C", 1:30, sep="")
si.gp <- factor(c(rep("E", 15), rep("M", 15)), levels=c("E", "M"))
stats_chisq.test(dat, si.gp, rowlevel=c("0", "1"))
}
\seealso{
\code{\link{chisq.test}} which this function wraps.
}

