% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modelTesting.R
\name{stats_wilcox.test}
\alias{stats_wilcox.test}
\title{Performs Wilcoxon tests between each row in a matrix and a vector using `stats` package in R}
\usage{
stats_wilcox.test(data, group, gpMat = FALSE, ...)
}
\arguments{
\item{data}{a matrix with at least 2 outcomes (0/1; Mut/WT, etc).}

\item{group}{a vector of factor which will be concatenated to the row of 'data'.}

\item{rowlevel}{specifies the order of factor use in a contingency table.}
}
\value{
the contingency table and p-value.
}
\description{
'Wilcox.test' performs one- and two-sample Wilcoxon tests on vectors of data; the latter
 is also known as 'Mann-Whitney' test.
}
\examples{
dat <- matrix(rnorm(10000*15*2), 5, 30)
rownames(dat) <- paste("R", 1:5, sep="")
colnames(dat) <- paste("C", 1:30, sep="")
si.gp <- factor(c(rep("E", 15), rep("M", 15)), levels=c("E", "M"))
stats_wilcox.test(dat, si.gp, gpMat=FALSE)
m0 <- matrix(0, 5, 30)
dat <- apply(m0, c(1,2), function(x) sample(c(0,1),1))
dat[1,] <- 0
dat[2,1] <- NA
rownames(dat) <- paste("R", 1:5, sep="")
colnames(dat) <- paste("C", 1:30, sep="")
si.gp <- rchisq(30,2)
stats_wilcox.test(dat, si.gp, gpMat=TRUE)
}
\seealso{
\code{\link{chisq.test}} which this function wraps.
}

