% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{webtable}
\alias{webtable}
\title{Show the data in a html page}
\usage{
webtable(...)
}
\arguments{
\item{...}{further parameters to \code{"DT::datatable"}.}
}
\description{
Show the data in a html page
}
