% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_plots.R
\name{bbplt}
\alias{bbplt}
\title{Boxplots or barplots from matrix with unique gene symbols}
\usage{
bbplt(
  g,
  mtx,
  s_meta,
  cond,
  plot_type = c("boxplot", "violinplot", "barplot"),
  plt_palette,
  with_beeswarm = FALSE,
  test_comp = list(),
  pre_calulated_test = NULL,
  test_method = "t.test",
  title = NULL,
  subtitle = NULL,
  title_wrap = 50,
  xlab = "",
  ylab = "",
  ylim = NULL,
  ylab_rotate = FALSE,
  paired_test = FALSE,
  return_plt = FALSE,
  width = 6,
  height = 7,
  plt_pref = "",
  plt_dir = "."
)
}
\arguments{
\item{g, }{a gene to be plotted}

\item{mtx, }{genes by samples matrix, \code{"g"} used rownames.}

\item{s_meta, }{meta info of samples with grouping info for ploting. There MUST be a "match_id" column where are colnames of mtx (orders don't matter).}

\item{cond, }{a column name to be used as grouping (e.g., x-axis).}

\item{plot_type, }{"boxplot" or "barplot".}

\item{plt_palette, }{color palette for plotting.}

\item{with_beeswarm, }{whether add beeswarm points to the boxplots.}

\item{test_comp, }{list for comparisons among groups. E.g., list(c("A", "B")): A vs. B. Student's t-test is used.}

\item{pre_calulated_test, }{a data.table MUST with columns "symbol" (unique) and "pvalue".}

\item{test_method, }{t-test is used as default.}

\item{title, }{title info.}

\item{subtitle, }{subtitle info.}

\item{title_wrap, }{number of characters to wrap the title.}

\item{xlab, }{xlab info.}

\item{ylab, }{ylab info.}

\item{ylim, }{a vector of y limits. c(NA, 1): NA means automatic scaling.}

\item{ylab_rotate, }{whether rotate the ylab 90 degree.}

\item{paired_test, }{whether paired t-test is used.}

\item{return_plt, }{whether return the ggplot object.}

\item{width, }{width of the figure.}

\item{height, }{height of the figure.}

\item{plt_pref, }{prefix for the figure names.}

\item{plt_dir, }{folder to save the figure.}
}
\value{
Figures in the sepecified folder.
}
\description{
Draw barplots or boxplots for matrix with unique gene symbols, such as matrix use symbols as rownames.
}
