% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplt_mult}
\alias{boxplt_mult}
\title{Boxplots from matrix with NON-unique gene symbols}
\usage{
boxplt_mult(x, is_unique = FALSE, gene_anno, mtx, ...)
}
\arguments{
\item{x, }{a gene symbol or a probe/ensembl id.}

\item{is_unique, }{is \code{"x"} representing a unique id, e.g., proble id.}

\item{gene_anno, }{match between ids and symbols, MUST have "uniq_id", and "symbol" columns.}

\item{mtx, }{gene matrix, ids by symbols.}

\item{..., }{parameters to}
}
\value{
Save the figures in the sepecified folder.
}
\description{
Draw boxplots for matrix with unique gene symbols, such as microaray data or matrix use ensembl ids as rownames.
Use boxplt_uniq_mult internally.
}
