% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_plots.R
\name{boxplt_uniq_mult}
\alias{boxplt_uniq_mult}
\title{Boxplots from matrix with unique ids}
\usage{
boxplt_uniq_mult(
  one_id,
  gene_anno,
  mtx,
  s_meta,
  cond,
  plt_palette,
  test_comp = list(),
  test_method = "t.test",
  paired_test = FALSE,
  title,
  subtitle = NULL,
  xlab = "",
  ylab = "",
  width = 6,
  height = 7,
  plt_pref = "",
  plt_dir = "."
)
}
\arguments{
\item{one_id, }{a single unique id which is rownames of the input.}

\item{gene_anno, }{match between ids and symbols, MUST have "uniq_id", and "symbol" columns.}

\item{mtx, }{input matrix, row names contains \code{"one_id"}.}

\item{s_meta, }{meta info of samples with grouping info for plotting. There MUST be a "match_id" column where are colnames of mtx (orders don't matter).}

\item{cond, }{a column name to be used as grouping (e.g., x-axis).}

\item{plt_palette, }{color palette for plotting.}

\item{test_comp, }{list for comparisons among groups. E.g., list(c("A", "B")): A vs. B. Student's t-test is used.}

\item{test_method, }{t-test is used as default.}

\item{paired_test, }{whether paired t-test is used.}

\item{title, }{title info.}

\item{subtitle, }{subtitle info.}

\item{xlab, }{xlab info.}

\item{ylab, }{ylab info.}

\item{width, }{width of the figure.}

\item{height, }{height of the figure.}

\item{plt_pref, }{prefix for the figure names.}

\item{plt_dir, }{folder to save the figure.}
}
\value{
Save the figures in the sepecified folder.
}
\description{
Draw boxplots for matrix with unique ids but not symbols, such as microaray data or matrix use ensembl ids as rownames.
}
