% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{depmap_cosmic_mut_cna}
\alias{depmap_cosmic_mut_cna}
\title{Collect and merge non-synonymous mutations and CNA from Depmap and cosmic for cell lines.}
\usage{
depmap_cosmic_mut_cna(
  g,
  sample_ano,
  dp_cm,
  dp_expr,
  MUT_CNA = FALSE,
  dp_mut,
  dp_cna,
  cm_mut,
  cm_cna
)
}
\arguments{
\item{g, }{a vector of interested genes. !! First check if any symbols missed in the Depmap expression data.}

\item{sample_ano, }{a data.table MUST with columns of cell_line and ModelID. If ModelID is not available, then it is "".}

\item{dp_cm, }{cosmic id <--> Depmap id.}

\item{dp_expr, }{Depmap expression data.}

\item{MUT_CNA, }{whether extracting MUT and CNA data from Depmap and Cosmic.}

\item{dp_mut, }{Depmap mutations data.}

\item{dp_cna, }{Depmap CNA data.}

\item{cm_mut, }{Cosmic mutations data.}

\item{cm_cna, }{Cosmic CNA data.}
}
\value{
A list of annotations.
}
\description{
Get the annoations of cell lines relying on the Depmap IDs.
}
\examples{
\dontrun{
load("Depmap/24Q2/data.DepMap.24Q2.rda")
#- Get the symbols for CNA and expression.
colnames(cna_tab)  <- gsub("\\\\s.*$", "", colnames(cna_tab), perl = TRUE)
colnames(expr_tab) <- gsub("\\\\s.*$", "", colnames(expr_tab), perl = TRUE)
#- COSMIC data.
cm_mut <- qs::qread("CellLinesProject_GenomeScreensMutant_v100_GRCh38.qs", nthreads = 4)
cm_cna <- readRDS("CellLinesProject_CompleteCNA_v100_GRCh38.rds")
dp_cm  <- readRDS("CellLinesProject_Sample_v100_GRCh38_depmapID.rds")

#- Mutations and CNA
g <- c("KRAS", "TP53", "ALK", "BRAF", "CDKN2A")
res_1 <- depmap_cosmic_mut_cna(g, ano, dp_cm, expr_tab, MUT_CNA = TRUE,
 dp_mut_tab, cna_tab, cm_mut, cm_cna)

#- Expression for the CBM genes.
g <- c("KRAS", "EGFR", "CARD9", "CARD10")
res_2 <- depmap_cosmic_mut_cna(g, ano, dp_cm, expr_tab, MUT_CNA = FALSE)
}
}
