% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{depmap_cosmic_mut_cna}
\alias{depmap_cosmic_mut_cna}
\title{Collect and merge non-synonymous mutations and CNA from Depmap and cosmic for cell lines.}
\usage{
depmap_cosmic_mut_cna(
  g,
  sample_ano,
  dp_cm,
  dp_expr,
  MUT_CNA = FALSE,
  dp_mut,
  dp_cna,
  cm_mut,
  cm_cna
)
}
\arguments{
\item{g, }{a vector of interested genes. !! First check if any symbols missed in the Depmap expression data.}

\item{sample_ano, }{a data.table MUST with columns of cell_line and ModelID. If ModelID is not available, then it is "".}

\item{dp_cm, }{cosmic id <--> Depmap id.}

\item{dp_expr, }{Depmap expression data.}

\item{MUT_CNA, }{whether extracting MUT and CNA data from Depmap and Cosmic.}

\item{dp_mut, }{Depmap mutations data.}

\item{dp_cna, }{Depmap CNA data.}

\item{cm_mut, }{Cosmic mutations data.}

\item{cm_cna, }{Cosmic CNA data.}
}
\value{
A list of annotations.
}
\description{
Get the annoations of cell lines relying on the Depmap IDs.
}
