% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{DiscretePalette}
\alias{DiscretePalette}
\title{Discrete colour palettes from the pals package}
\usage{
DiscretePalette(n, palette = NULL)
}
\arguments{
\item{n}{Number of colours to be generated.}

\item{palette}{Options are
"alphabet" (26), "alphabet2" (26), "glasbey" (32), "polychrome" (36), and "stepped" (24).
Can be omitted and the function will use the one based on the requested n.}
}
\value{
A vector of colors
}
\description{
Taken from Seurat, which taken from pals.
}
