% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_david.R
\name{flat_david}
\alias{flat_david}
\title{Flatten DAVID GO annotation}
\usage{
flat_david(GOres, p_thres = 0.001, fdr_thres = NULL, anno = NULL)
}
\arguments{
\item{GOres}{a data.table with DAVID GO annotation.}

\item{p_thres}{threshold for pvalue, only used if fdr_thres is NULL.}

\item{fdr_thres}{threshold for FDR.}

\item{anno}{a data.table with two columns: "enid" and "symbol".}
}
\value{
A data.table with flatted DAVID results, one gene per line.
}
\description{
Conver the summary DAVID results to one gene per line, filtered by Pvalue or FDR.
}
