% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{flat_GSEA}
\alias{flat_GSEA}
\title{Flat and save GSEA annotation generated by fgsea.}
\usage{
flat_GSEA(
  x,
  db,
  prefix_name,
  proj_name,
  output_dir = "./gsea_plot",
  keep_all = c("h", "kegg"),
  width = 7,
  height = 5,
  save_png = TRUE
)
}
\arguments{
\item{x}{the input list, with names like c("input_gene", "h", "c2").}

\item{db}{GSEA annotation database, MUST be in the same order of elements in x.}

\item{prefix_name}{names used for plot folders, normal the comparisons.}

\item{proj_name}{data projects.}

\item{output_dir}{dir to save outputs and plots.}

\item{keep_all}{keep all gene sets for sepecified groups.}

\item{width}{width of figures.}

\item{height}{height of figures}

\item{save_png}{boolean, TRUE to save figures to png files, FALSE to save in pdf files.}
}
\value{
A vector of colors
}
\description{
Taken a list with different GSEA category annotation, save the results and plots.
}
