% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gde}
\alias{gde}
\title{Extract the differentially expressed genes}
\usage{
gde(dds, contrast, ganno)
}
\arguments{
\item{dds}{a DESeq2 returned results. The input matrix is expected to have the format of
genes by samples, while genes are ensembl gene ids.}

\item{contrast}{contrast to be used.}

\item{ganno}{a data.table containing genes annotation, must have two columns with
"geneid", and "symbol".}
}
\value{
A data.table with all results.
}
\description{
Get the DE genes from DESeq2 results, removing the NA rows.
}
\examples{
\dontrun{
ll::gde(dds, contrast = c("cond", "CHARGE", "WT"), ganno = unique(ll::mm10enid[, -"transid"]))
}
}
