% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19enid}
\alias{hg19enid}
\title{human gene annotation}
\format{
A data.table with three columns: "geneid","transid", "symbol", "uniquesymbol".
}
\usage{
hg19enid
}
\description{
Ensembl hg19 gene annotation (v.75)
}
\keyword{datasets}
