% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10enid}
\alias{mm10enid}
\title{mouse gene annotation}
\format{
A data.table with three columns: "geneid","transid", "symbol", "uniquesymbol".
}
\usage{
mm10enid
}
\description{
Ensembl mm10 gene annotation (v.86)
}
\keyword{datasets}
