% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_godsnot_102}
\alias{palette_godsnot_102}
\title{scvelo 102 color}
\format{
A color palette extracted from scvelo, with #FEFFE6 removed as too light.
}
\usage{
palette_godsnot_102
}
\description{
scvelo 102 color
}
\keyword{datasets}
