% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsplit}
\alias{rsplit}
\title{Split a string from right end}
\usage{
rsplit(x, s, n = Inf)
}
\arguments{
\item{x}{a string to be splitted.}

\item{s}{the separator, "fixed".}

\item{n}{number of desired pieces.}
}
\value{
A matrix with splitted parts.
}
\description{
Split a string from the righ end and return the result vector. The number of splits could
be specified.
}
\examples{
\dontrun{
rsplit("A_B_C_D_E", "_", 3)
}
}
