% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_data}
\alias{scale_data}
\title{Scale data to arbitrary range}
\usage{
scale_data(x, lb = 0, ub = 1)
}
\arguments{
\item{x}{numeric vector.}

\item{lb}{low boundary, 0 by default}

\item{ub}{up boundary, 1 by default.}
}
\value{
A numeric vector.
}
\description{
Scale vector to sepecified range, by default [0, 1]
}
