% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{split_updown}
\alias{split_updown}
\title{Get the ids of down/up-regulated genes from DESeq2.}
\usage{
split_updown(x, to_file = TRUE)
}
\arguments{
\item{x}{path to a file, which must have two columns of "log2FoldChange" and "geneid"}

\item{to_file}{whether save up-/down-regulated files.}
}
\value{
A list gene ids of up-/down-regulated genes. Related files are writen to the same folder.
}
\description{
Get the ids of down/up-regulated genes from DESeq2.
}
