% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao_utils.R
\name{tc_cluster}
\alias{tc_cluster}
\title{Cluster the time-course data by K-means/PAM}
\usage{
tc_cluster(
  indata,
  ncl,
  timepoint,
  cluster_method = c("kmeans", "pam"),
  dist_method = function(x) {
     as.dist((1 - cor(t(x)))/2)
 },
  xlab_name,
  ylab_name,
  save_fig = TRUE,
  width = 12,
  height = 10,
  clustercolour = rainbow(20),
  figure_prefix = "",
  show_se = FALSE
)
}
\arguments{
\item{indata}{a matrix with gene in rows and time points in columns. Colnames have the pattern of "T_1", "T_2", etc.}

\item{ncl}{number of k-means clusters.}

\item{timepoint}{a data table with two column t1 and t2, t1 is the character name of time, and t2 is the numberic time.}

\item{cluster_method}{K-means or PAM.}

\item{dist_method}{method to caluclate the metric used in PAM.}

\item{xlab_name}{name for the xlab.}

\item{ylab_name}{name for the ylab.}

\item{save_fig}{whether saving the plot.}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{clustercolour}{color themes for the smooth line of kmeans clusters.}

\item{figure_prefix}{prefix for the figure}

\item{show_se}{whether show the se of the gam smooth line.}
}
\value{
A ggplot object and a csv file with gene clustering information saved in the working folder.
}
\description{
Perform and plot the K-means or PAM clustering on time course data.
}
